Refactorisation in accordance with the MVC architecture
