+++ Phonebook/src/main/java/application/main/Model/Database/DAOs/ContactInfoDAO.java
@@ -1,8 +1,8 @@
-package application.main.Database.DAOs;
+package application.main.Model.Database.DAOs;
 
-import application.main.Database.DatabaseHandlerFactory;
-import application.main.Database.Interfaces.IContactInfoDAO;
-import application.main.Entities.ContactInfo;
+import application.main.Model.Database.DatabaseHandlerFactory;
+import application.main.Model.Database.Interfaces.IContactInfoDAO;
+import application.main.Model.Entity.ContactInfo;
 import com.microsoft.sqlserver.jdbc.SQLServerDriver;
 
 import java.sql.*;
