+++ Phonebook/src/main/java/application/main/Model/Entity/Address.java
@@ -1,4 +1,4 @@
-package application.main.Entities;
+package application.main.Model.Entity;
 
 import java.util.ArrayList;
