+++ Phonebook/src/main/java/application/main/Model/Entity/ContactInfo.java
@@ -1,4 +1,4 @@
-package application.main.Entities;
+package application.main.Model.Entity;
 
 public class ContactInfo
 {
