+++ Phonebook/src/test/java/AddressTest.java
@@ -1,5 +1,5 @@
-import application.main.Entities.Address;
-import application.main.Entities.ContactInfo;
+import application.main.Model.Entity.Address;
+import application.main.Model.Entity.ContactInfo;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
